generate_wave {rearrr}	R Documentation

Generate simple wave signal

Description

Generate y-values for the wave signal at the given time points with:

amplitude * trig_fn(2 * pi * (1 / fs) * ts + phase)

Usage

generate_wave(ts, fs = 44100, amplitude = 1, phase = 0, trig_fn = sin)

generate_sine_wave(ts, fs = 44100, amplitude = 1, phase = 0)

generate_cosine_wave(ts, fs = 44100, amplitude = 1, phase = 0)

Arguments

ts
Time points.

fs
Sampling frequency (samples per second).

Can be a vector with one frequency per time point.

amplitude
Scaling constant.

Can be a vector with one amplitude per time point.

phase
Offset in radians.

Can be a vector with one phase offset per time point.

trig_fn
Trigonometric function like sin and cos.

Value

y-values for the wave signal.

[Package rearrr version 0.3.5 Index]